Rename as blocks/CurrentSchoolBlock
